image opening functions fixed, textures.c created
